Logout moved to its own component and Header component added. Logout and header moved from home.js. Sign out still visible on login page!
