+++ frontend/src/App.js
@@ -1,6 +1,7 @@
 import Login from './routes/Login'
 import Home from './routes/Home'
 import Orgs from './routes/Orgs'
+import Header from './components/Header'
 import React from 'react'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 
@@ -8,6 +9,7 @@ function App() {
 
   return (
     <Router>
+      <Header/>
       <Routes>
         <Route path="/login" element={<Login />} />
         <Route path="/" element={<Home />} />
