+++ frontend/src/routes/Home.js
@@ -1,14 +1,9 @@
 import { useEffect } from 'react'
-import { Navigate, useNavigate, Link } from 'react-router-dom'
+import { useNavigate, Link } from 'react-router-dom'
 import authService from '../services/auth'
 
 const Home = () => {
   const navigate = useNavigate()
-
-  const logout = () => {
-    localStorage.removeItem('token')
-    return <Navigate to="/login" replace/>
-  }
 
   const checkAuth = async () => {
     const token = localStorage.getItem('token')
@@ -28,12 +23,6 @@ const Home = () => {
 
   return (
     <div>
-      <div className="flex justify-between items-center p-2 bg-blue-600">
-        <h2 className="text-white">TMC Coach</h2>
-        <form onSubmit={logout} className="bg-red-500 w-fit px-2.5 py-2 rounded text-white">
-          <button type="submit">Sign out</button>
-        </form>
-      </div>
       <br></br>
       <br></br>
       <Link to="/orgs">Organizations</Link>
